Deprecate scale scheduler algorithm and move to trainer
