+++ composer/trainer/trainer.py
@@ -19,6 +19,8 @@ from torch.optim.lr_scheduler import CosineAnnealingLR
 from torchmetrics.collections import MetricCollection
 from torchmetrics.metric import Metric
 
+from composer.composer.algorithms.scale_schedule.scale_schedule import ScaleSchedule
+from composer.composer.trainer.scale_schedule import scale_scheduler
 from composer.core import Callback, DataSpec, Engine, Event, Logger, State, Time, surgery
 from composer.core.algorithm import Algorithm
 from composer.core.evaluator import Evaluator
@@ -150,6 +152,7 @@ class Trainer:
         validate_every_n_epochs: int = 1,
         compute_training_metrics: bool = False,
         precision: Union[str, Precision] = Precision.FP32,
+        scale_schedule_ratio: float = 1.0,
 
         # dist hparams
         dist_timeout: float = 300.0,
@@ -190,8 +193,23 @@ class Trainer:
         # self._use_grad_scaling() will raise a RuntimeError if grad scaling is not available when it is required
         warnings.filterwarnings(action="ignore", message="torch.cuda.amp.GradScaler")
 
+        # ScaleSchedule is a deprecated algorithm, but if it is used, updated SSR with its ratio.
+        # TODO(#434): Remove this completely.
+        for algorithm in algorithms or []:
+            if isinstance(algorithm, ScaleSchedule):
+                scale_schedule_ratio = algorithm.ratio
+
         if isinstance(max_duration, str):
             max_duration = Time.from_timestring(max_duration)
+
+        orig_max_duration = max_duration
+
+        if scale_schedule_ratio != 1.0:
+            max_duration = cast(Time[int], orig_max_duration * scale_schedule_ratio)
+            log.info(f'max_duration changed from {orig_max_duration} to {max_duration}')
+            if max_duration.value == 0:
+                raise ValueError(
+                    'Scale schedule has reduced the max_duration to 0. Set a higher ratio or use more epochs.')
 
         self.deepspeed_config = deepspeed_config
 
@@ -297,6 +315,13 @@ class Trainer:
                 raise ValueError("If a scheduler is not provided, max duration must be in epochs")
             schedulers = CosineAnnealingLR(optimizer, T_max=max_duration.value)
             warnings.warn(f"No scheduler was specified. Defaulting to {repr(schedulers)}")
+
+        if scale_schedule_ratio != 1.0:
+            if orig_max_duration.unit != TimeUnit.EPOCH:
+                raise NotImplementedError("Max duration must be specified in epochs. Other units are not yet supported.")
+
+            for scheduler in ensure_tuple(schedulers):
+                scale_scheduler(scheduler, scale_schedule_ratio, orig_max_duration.value)
 
         schedulers = ComposedScheduler(ensure_tuple(schedulers))
